pyproject.toml
[tool.poetry]
name = "appbackend"
version = "0.1.0"
description = "Template for fastapi backend"
authors = ["Shri <shrinidhi.hegde@equentis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
alembic = "^1.10.4"
psycopg2 = "^2.9.6"
httpx = "^0.24.0"
pytest-asyncio = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
